requirements.txt
rasterio
numpy
docker
pillow
tqdm
opencv-python